Deleted non-compiling test with original printf and replace cc by clang according to 42 school

+++ src/test_upperx.c
@@ -6,7 +6,7 @@
 /*   By: thugueno <thugueno@student.42angoulem      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/25 17:32:41 by thugueno          #+#    #+#             */
-/*   Updated: 2022/10/26 01:14:36 by thugueno         ###   ########.fr       */
+/*   Updated: 2022/10/31 15:26:05 by thugueno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ void	test_upperx(void)
 	TEST(5, "%X %X", 90, -90);
 	TEST(6, " %X ", INT_MIN);
 	TEST(7, " %X ", INT_MAX);
-	TEST(8, " %X ", LONG_MAX);
-	TEST(9, " %X ", LONG_MIN);
+	TEST(8, " %X ", (unsigned int)LONG_MAX);
+	TEST(9, " %X ", (unsigned int)LONG_MIN);
 	end_category();
 }
